+++ spec/component_guide/component_example_spec.rb
@@ -32,14 +32,12 @@ describe "Component example" do
     visit "/component-guide/test-component"
 
     expect(page).to have_selector('link[href*="/assets/application.css"]', visible: false)
-    expect(page).to have_selector('link[href*="/assets/print.css"]', visible: false)
   end
 
   it "doesn't include the application stylesheet for a gem component" do
     visit "/component-guide/button"
 
     expect(page).not_to have_selector('link[href*="/assets/application.css"]', visible: false)
-    expect(page).not_to have_selector('link[href*="/assets/print.css"]', visible: false)
   end
 
   it "lists examples in a human readable way" do