+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
+## Unreleased
+
+* **BREAKING** Remove print stylesheets from components ([PR #3110](https://github.com/alphagov/govuk_publishing_components/pull/3110))
+
 ## 33.1.0
 
 * Update feedback component tracking ([PR #3099](https://github.com/alphagov/govuk_publishing_components/pull/3099))